Moved Query and Accessor-interfaces -> Language ns
