+++ src/Storage/Database/Language/SQL/Query/Select.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Projom\Storage\Database\Language\SQL\Query;
 
+use Projom\Storage\Database\Language\QueryInterface;
 use Projom\Storage\Database\Language\SQL\Column;
 use Projom\Storage\Database\Language\SQL\Filter;
 use Projom\Storage\Database\Language\SQL\Group;
@@ -11,9 +12,8 @@ use Projom\Storage\Database\Language\SQL\Limit;
 use Projom\Storage\Database\Language\SQL\Order;
 use Projom\Storage\Database\Language\SQL\Table;
 use Projom\Storage\Database\Language\SQL\Join;
-use Projom\Storage\Database\Language\SQL\QueryInterface;
-use Projom\Storage\Database\Language\SQL\Util;
 use Projom\Storage\Database\Language\SQL\QueryObject;
+use Projom\Storage\Database\Language\SQL\Util;
 
 class Select implements QueryInterface
 {
